Adds mechanism to reset count and others
Adds mechanism to reset count - using open cv's waitKey() function to
get a keypress and comparing it against the ASCII values of Q,q,R,r to
either break the loop (Q,q) or reseting the value of count (r,R).

Also, previosuly a simple check of previous_pose against '' would
produce warning which is prevented by checking it in another novel way -
checking whether the variable is a string.

Further the default value of tolerance has been tweaked to 40. Expect
the value to undergo large level tunings in future commit messages as
well, the exact value is yet to be determined.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>